explicitly install httparty gem at an older revision
fauxhai (referenced through searchef) does not pin httparty, and a
recent release of httparty introduces a dependency on json-1.8.
Unfortunately chef requires 1.4-1.7.  By explicitly installing an
earlier version of httparty we can avoid this for now.

Longer term we'll need to address this in fauxhai upstream, assuming we
can't update chef itself for compatibility with json-1.8